Moving AlwaysNotTakenPredictor files to BP directory
