+++ src/include/simeng/CoreInstance.hh
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "simeng/AlwaysNotTakenPredictor.hh"
+#include "simeng/branchPredictors/AlwaysNotTakenPredictor.hh"
 #include "simeng/Core.hh"
 #include "simeng/Elf.hh"
 #include "simeng/SpecialFileDirGen.hh"
